+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * upgraded to Spring Boot 3.0.0 (not backwards compatible with previous versions of Spring Data)
 * switched over to Jakarta dependencies with jakarta packages (e.g. jakarta.persistence-api 3.0.0)
+* replaced use of `PersistenceConstructor` with `PersistenceCreator` (`PersistenceConstructor` is no longer supported!)
 
 ### 7.2.0
 * added basic support for `@MappedCollections`